Added barcode scanner page and made it work for the most part Added Custom Medicine add, edit, and list page and manager and made them work for the most part Added custom medicine saving and loading Added new button to go to barcode scanner Added menu items for custom medicine list and to add a custom medicine Added recycler adapter for custom medicine list page Added a bunch of layouts and tweaked a few others

+++ app/src/main/java/com/example/medicinetrackingapp/IndividualMedicine.java
@@ -2,13 +2,12 @@ package com.example.medicinetrackingapp;
 
 import java.io.Serializable;
 import java.util.Calendar;
-import java.util.Date;
 
 public class IndividualMedicine implements Serializable, Comparable<IndividualMedicine> {
     public String name;
     public Calendar takenDateTime;
     public final Calendar inputTimeDate;
-    public Long quantity;
+    public int quantity;
     public String reason;
     public int dose;
